Removed warnings from clib at level 4. Just two remain: one is related to stdafx.obj, the other to structured exceptions. Still need to investigate those.

+++ pol-core/clib/threadhelp.h
@@ -89,7 +89,7 @@ namespace Pol {
       ~DynTaskThreadPool();
       void push( msg msg );
       std::future<bool> checked_push( msg msg );
-      unsigned int threadpoolsize() const;
+      size_t threadpoolsize() const;
     protected:
       bool _done;
     private:
